20:02 06.12.2018 v

+++ sandbox.controller/src/kz/greetgo/sandbox/controller/model/Charm.java
@@ -4,5 +4,5 @@ public class Charm {
     public int id;
     public String name;
     public String description;
-    public float energy;
+    public double energy;
 }
